Move to ES6 - WIP
