+++ Main/template-settings.js
@@ -1,3 +1,4 @@
+export default 
 {
     "mode": "production",
     "connections": {
