+++ deploy-commands.js
@@ -2,7 +2,7 @@ import { REST } from '@discordjs/rest';
 import { Routes } from 'discord-api-types/v10';
 import { Collection } from 'discord.js';
 
-const settings = require('./Main/settings.json');
+import settings from './Main/settings.js';
 
 let token;
 let bot = {};
@@ -10,11 +10,9 @@ bot.value = 'value';
 bot.commands = new Collection();
 bot.commandsArray = [];
 
-require('./Main/Handlers/commands.js')(bot);
-// console.log(bot.commandsArray);
+import { commandHandler } from './Main/Handlers/commands.js';
+commandHandler(bot);
 
-
-// let commandsArray = JSON.parse(bot.commandsArray);
 let commandsArray = bot.commandsArray;
 
 //Check if in production or development
